==> Part_1.ipynb <==
import pandas as pd
import seaborn as sns
import numpy as np
import matplotlib.pyplot as plt

basics_url="title.basics.tsv.gz"
akas_url="title.akas.tsv.gz"
ratings_url="title.ratings.tsv.gz"

#Load file
basics_df = pd.read_csv(basics_url, sep='\t', low_memory=False)
ratings = pd.read_csv(ratings_url, sep='\t', low_memory=False)
akas_df = pd.read_csv(akas_url, sep='\t', low_memory=False)

basics_df.info()

akas_df.info()

basics_df = basics_df.replace({'\\N':np.nan})
akas_df = akas_df.replace({'\\N':np.nan})
basics_df.head()




akas_df.head()

# Exclude any movie with missing values for genre or runtime
is_missing_genres = basics_df['genres'].isna()
basics_df = basics_df[~is_missing_genres]
basics_df

is_missing_runtime = basics_df['runtimeMinutes'].isna()
basics_df = basics_df[~is_missing_runtime]
basics_df

# Exclude movies that are included in the documentary category.
is_documentary =  basics_df['genres'].str.contains('documentary',case=False)
basics_df =  basics_df[~is_documentary]
basics_df.head()

# Filter the basics table down to only include the US by using the filter akas dataframe
keepers =basics_df['tconst'].isin(akas_df['titleId'])
basics_df = basics_df[keepers]
basics_df

# Calculate the mean for Item_Weight column
is_missing_startYear = basics_df['startYear'].isna()
basics_df = basics_df[~is_missing_startYear]
basics_df

basics_df = basics_df.loc[(basics_df['startYear'] >= '2000') & (basics_df['startYear'] <= '2021')]
basics_df

# example making new folder with os
import os
os.makedirs('Data/',exist_ok=True) 
# Confirm folder created
os.listdir("Data/")

## Save current dataframe to file.
basics_df.to_csv("Data/title_basics.csv.gz",compression='gzip',index=False)

# Open saved file and preview again
basics_df = pd.read_csv("Data/title_basics.csv.gz", low_memory = False)
basics_df.head()



## Save current dataframe to file.
akas_df.to_csv("Data/title_aka.csv.gz",compression='gzip',index=False)
# Open saved file and preview again
akas_df = pd.read_csv("Data/title_akas.csv.gz", low_memory = False)
akas_df.head()


## Save current dataframe to file.
ratings.to_csv("Data/title_ratings.csv.gz",compression='gzip',index=False)
# Open saved file and preview again
ratings = pd.read_csv("Data/title_ratings.csv.gz", low_memory = False)
ratings.head()
