==> Part_3.ipynb <==
# Import relevent liberies 
import pandas as pd 
import numpy as ny 


# Loading basic-rating
basic = pd.read_csv('Data/title_basics.csv.gz')
basic.head()

## create a col with a list of genres
basic['genres_split'] = basic['genres'].str.split(',')
basic['genres_split']


#Use explode function
exploded_genres = basic.explode('genres_split')
exploded_genres


unique_genres = sorted(exploded_genres['genres_split'].unique())
unique_genres

# Creating a new table from Genres
title_genres = exploded_genres[['tconst','genres_split']].copy()
title_genres.head()

## Making the genre mapper dictionary
genre_ints = range(len(unique_genres))
genre_map = dict(zip(unique_genres, genre_ints))
genre_map


genre_id_map = dict(zip(unique_genres,range(len(unique_genres))))
genre_id_map

## make new integer genre_id and drop string genres
title_genres['genre_id'] = title_genres['genres_split'].map(genre_map)
title_genres = title_genres.drop(columns='genres_split')


title_genres.head()

genre_lookup =pd.DataFrame({'genre_name':genre_id_map.keys(),
                            'genre_Id' :genre_id_map.values()})
genre_lookup

basic.head()

basic = basic[['tconst','primaryTitle','startYear','endYear','runtimeMinutes']]
basic.head()

ratings = pd.read_csv('Data/title_ratings.csv.gz')
ratings.head()

# loading `TMBD Api Result 
tmdb_Api = pd.read_csv('Data/Datatmdb_results_combined.csv.gz')
tmdb_Api.info()

tmdb_Api =tmdb_Api[['imdb_id','budget','revenue','certification']]
tmdb_Api.head()

# checking datatype tmdb_Api
tmdb_Api.dtypes

# Checking datatype for ratings
ratings.dtypes

# Checking datatype for basic
basic.dtypes

import pandas as pd
import pymysql
pymysql.install_as_MySQLdb()
from sqlalchemy import create_engine
from sqlalchemy_utils import create_database, database_exists

from urllib.parse import quote_plus as urlquote
from sqlalchemy.engine import create_engine

# Create connection string using credentials following this format
# connection = "dialect+driver://username:password@host:port/database"
username = 'root'
password = 'Lelozizi@1059' # (or whatever password you chose during mysql installation)
db_name = 'movies'

connection_str = f"mysql+pymysql://{'root'}:{urlquote('Lelozizi@1059**')}@localhost/{db_name}"


engine = create_engine(connection_str)
engine

# Check if the database exists. If not, create it.
if database_exists(connection_str) == False:
  create_database(connection_str)
else:
  print('The database already exists')


basic.head()

## load
from sqlalchemy.types import *
## Calculate max string lengths for object columns
key_len = basic['tconst'].fillna('').map(len).max()
title_len = basic['primaryTitle'].fillna('').map(len).max()
## Create a schema dictonary using Sqlalchemy datatype objects
df_schema = {
    "tconst": String(key_len+1), 
    "primaryTitle": Text(title_len+1),
    'startYear':Float(),
    'endYear':Float(),
    'runtimeMinutes':Integer()}


# Save to sql with dtype and index=False
basic.to_sql('title_basics',engine,dtype=df_schema,if_exists='replace',index=False)


ratings schema 

# creating schema for rating 
key_len = ratings['tconst'].fillna('').map(len).max()
## Create a schema dictonary using Sqlalchemy datatype objects
df_schemarating = {
    "tconst": String(key_len+1), 
    'averageRating':Float(),
    'numVotes':Integer()}

# Save to sql with dtype and index=False
ratings.to_sql('title_ratings',engine,dtype=df_schemarating,if_exists='replace',index=False)

tmdb_Api schema

# creating schema for tmdb_Api
key_len = tmdb_Api['imdb_id'].fillna('').map(len).max()
cert_len = tmdb_Api['certification'].fillna('').map(len).max()
## Create a schema dictonary using Sqlalchemy datatype objects
df_schema = {
    "imdb_id": String(key_len+1), 
    "budget": Float(),
    'revenue':Float(),
    'certification':Text(cert_len+1)}

# Save to sql with dtype and index=False
tmdb_Api.to_sql('tmdb_data',engine,dtype=df_schema,if_exists='replace',index=False)

# Checking genre dttype 
genre_lookup.dtypes

# creating schema for genre
name_len = genre_lookup['genre_name'].fillna('').map(len).max()
## Create a schema dictonary using Sqlalchemy datatype objects
df_schemagenre = {
    "genre_Id": Integer(), 
    'genre_name':Text(name_len+1)}

# Save to sql with dtype and index=False
genre_lookup.to_sql('genra',engine,dtype=df_schemagenre,if_exists='replace',index=False)

#engine.execute('ALTER TABLE movies.genra ADD PRIMARY KEY (`genre_Id`);')


title_genres.dtypes

Title_Genre Schema 

# creating schema for title_genre
key_len = basic['tconst'].fillna('').map(len).max()
## Create a schema dictonary using Sqlalchemy datatype objects
df_schema = {
    "tconst": String(key_len+1), 
    'genre_id':Integer()}

# Save to sql with dtype and index=False
title_genres.to_sql('title_genres',engine,dtype=df_schema,if_exists='replace',index=False)

# show tables
q = """SHOW TABLES;"""
pd.read_sql(q, engine)

q1 = """SELECT * FROM movies.genra LIMIT 5;"""
pd.read_sql(q1, engine)


# Querying title_basics
q2 = """SELECT * FROM movies.title_basics LIMIT 5;"""
pd.read_sql(q2, engine)

#Querying title_genres

q3 = """SELECT * FROM movies.title_genres LIMIT 5;"""
pd.read_sql(q3, engine)

#Querying title_genres

q4 = """SELECT * FROM movies.title_ratings LIMIT 5;"""
pd.read_sql(q4, engine)

#Querying title_genres

q5 = """SELECT * FROM movies.tmdb_data LIMIT 5;"""
pd.read_sql(q5, engine)
